feat: create Theater to replace original Unit and Unit reused as parent.
