+++ public/css/stylesSelfFront.css
@@ -20,6 +20,69 @@
 ._C_C_app_Component_Draft_styleDisplay_module__markDisplayBlock div, ._C_C_app_Component_Draft_styleDisplay_module__markDisplayBlockNowrap div {
   display: inline-block;
 }
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxUnitLayerSwitch {
+  width: 100%;
+  height: 100%;
+  box-sizing: border-box;
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxSwitch {
+  width: 100%;
+  height: auto;
+  position: absolute;
+  top: 38%;
+  left: 50%;
+  transform: translate(-50%,0);
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxSwitchSvg {
+  height: auto;
+  position: relative;
+  z-index: 2;
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxSpots {
+  width: 100%;
+  height: 172%;
+  position: absolute;
+  box-sizing: border-box;
+  z-index: 0;
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxSpotBin {
+  width: 24%;
+  height: 75%;
+  position: absolute;
+  left: 50%;
+  transform: translate(-50%,0);
+  box-sizing: border-box;
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxSpotSing {
+  width: 100%;
+  height: auto;
+  position: absolute;
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxDnClick {
+  width: 100%;
+  height: 70%;
+  position: absolute;
+  bottom: 0;
+  cursor: pointer;
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxSwitUp {
+  width: 91%;
+  left: 50%;
+  transform: translate(-50%,0);
+  margin-bottom: 1.5%;
+  cursor: pointer;
+}
+
+._C_C_app_Unit_UnitLayerSwitch_styles_module__boxSwitDn {
+  width: 100%;
+}
 ._C_C_app_Component_Editing_EditingModal_styleEditingModal_module__imgBLockDecoBack {
   width: 18%;
   height: 100%;
@@ -60,70 +123,7 @@
   line-height: 1.7rem;
   color: #ABABAB;
 }
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxUnitLayerSwitch {
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxSwitch {
-  width: 100%;
-  height: auto;
-  position: absolute;
-  top: 38%;
-  left: 50%;
-  transform: translate(-50%,0);
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxSwitchSvg {
-  height: auto;
-  position: relative;
-  z-index: 2;
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxSpots {
-  width: 100%;
-  height: 172%;
-  position: absolute;
-  box-sizing: border-box;
-  z-index: 0;
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxSpotBin {
-  width: 24%;
-  height: 75%;
-  position: absolute;
-  left: 50%;
-  transform: translate(-50%,0);
-  box-sizing: border-box;
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxSpotSing {
-  width: 100%;
-  height: auto;
-  position: absolute;
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxDnClick {
-  width: 100%;
-  height: 70%;
-  position: absolute;
-  bottom: 0;
-  cursor: pointer;
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxSwitUp {
-  width: 91%;
-  left: 50%;
-  transform: translate(-50%,0);
-  margin-bottom: 1.5%;
-  cursor: pointer;
-}
-
-._C_C_app_Component_Unit_UnitLayerSwitch_styles_module__boxSwitDn {
-  width: 100%;
-}
-._C_C_app_Component_Unit_NodesDisplay_styles_module__boxListItem {
+._C_C_app_Unit_NodesDisplay_styles_module__boxListItem {
   display: inline-block;
   width: 100%;
   position: relative;
@@ -184,7 +184,7 @@
   word-break: break-all;
   color: #000000;
 }
-._C_C_app_Component_Unit_Mark_ViewerBlock_styles_module__boxInteraction {
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxInteraction {
   display: flex;
   flex-direction: column;
   flex-wrap: nowrap;
@@ -196,20 +196,20 @@
   background-color: #FCFCFC;
 }
 
-._C_C_app_Component_Unit_Mark_ViewerBlock_styles_module__boxPanelJump {
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxPanelJump {
   box-sizing: border-box;
   padding: 0 10% 0 8%;
   margin: 2.7vh 0 0 0;
   text-align: right;
 }
 
-._C_C_app_Component_Unit_Mark_ViewerBlock_styles_module__boxContentDraft {
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__boxContentDraft {
   min-height: 37vh;
   box-sizing: border-box;
   padding: 1vh 9% 0 10.8%;
 }
 
-._C_C_app_Component_Unit_Mark_ViewerBlock_styles_module__fontContentDraft {
+._C_C_app_Unit_Mark_AuthorBlock_styles_module__fontContentDraft {
   font-size: 1.52rem;
   letter-spacing: 0.07rem;
   line-height: 1.6rem;
@@ -218,7 +218,7 @@
   word-break: break-all;
   color: #000000;
 }
-._C_C_app_Component_Unit_Mark_AuthorBlock_styles_module__boxInteraction {
+._C_C_app_Unit_Mark_ViewerBlock_styles_module__boxInteraction {
   display: flex;
   flex-direction: column;
   flex-wrap: nowrap;
@@ -230,20 +230,20 @@
   background-color: #FCFCFC;
 }
 
-._C_C_app_Component_Unit_Mark_AuthorBlock_styles_module__boxPanelJump {
+._C_C_app_Unit_Mark_ViewerBlock_styles_module__boxPanelJump {
   box-sizing: border-box;
   padding: 0 10% 0 8%;
   margin: 2.7vh 0 0 0;
   text-align: right;
 }
 
-._C_C_app_Component_Unit_Mark_AuthorBlock_styles_module__boxContentDraft {
+._C_C_app_Unit_Mark_ViewerBlock_styles_module__boxContentDraft {
   min-height: 37vh;
   box-sizing: border-box;
   padding: 1vh 9% 0 10.8%;
 }
 
-._C_C_app_Component_Unit_Mark_AuthorBlock_styles_module__fontContentDraft {
+._C_C_app_Unit_Mark_ViewerBlock_styles_module__fontContentDraft {
   font-size: 1.52rem;
   letter-spacing: 0.07rem;
   line-height: 1.6rem;
